requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
astunparse==1.6.3
attrs==21.4.0
black==21.12b0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.26
dataclasses==0.6
fire==0.4.0
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hydra-core==1.1.1
idna==3.3
imageio==2.13.5
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
iopath==0.1.9
jsonschema==4.4.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelme2coco==0.2.1
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.5.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.5
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
pixellib==0.7.1
platformdirs==2.4.1
portalocker==2.3.2
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sahi==0.8.22
scikit-image==0.19.1
scipy==1.7.3
Shapely==1.8.0
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminaltables==3.1.10
tifffile==2021.11.2
tomli==1.2.3
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
tqdm==4.62.3
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
xmljson==0.2.1
yacs==0.1.8
zipp==3.7.0
